pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pomera-ai-commander"
dynamic = ["version"]
description = "Text processing toolkit with 22 MCP tools for AI assistants - case transformation, encoding, hashing, text analysis, and notes management"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "matbanik", email = "matbanik@users.noreply.github.com"}
]
maintainers = [
    {name = "matbanik", email = "matbanik@users.noreply.github.com"}
]
keywords = [
    "mcp",
    "model-context-protocol",
    "text-processing",
    "ai-tools",
    "pomera",
    "text-manipulation",
    "encoding",
    "hashing",
    "notes",
    "cursor",
    "claude-desktop"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing",
    "Topic :: Utilities"
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "platformdirs>=4.0.0",
    "deepdiff>=6.7.0,<9.0.0",
    "PyYAML>=6.0.0",
    "json5>=0.9.0",
    "jsonschema>=4.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
gui = [
    "reportlab>=3.6.0",
    "python-docx>=0.8.0",
    "rapidfuzz>=3.0.0",
]
full = [
    "reportlab>=3.6.0",
    "python-docx>=0.8.0",
    "rapidfuzz>=3.0.0",
    "pyaudio>=0.2.0",
    "numpy>=1.20.0",
    "psutil>=5.8.0",
    "huggingface-hub>=0.10.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/matbanik/Pomera-AI-Commander"
Documentation = "https://github.com/matbanik/Pomera-AI-Commander#readme"
Repository = "https://github.com/matbanik/Pomera-AI-Commander"
Issues = "https://github.com/matbanik/Pomera-AI-Commander/issues"
Changelog = "https://github.com/matbanik/Pomera-AI-Commander/releases"

[project.scripts]
pomera-ai-commander = "pomera_mcp_server:main"
pomera-mcp = "pomera_mcp_server:main"
pomera-create-shortcut = "create_shortcut:create_shortcut"

[tool.setuptools]
py-modules = ["pomera_mcp_server", "pomera", "create_shortcut"]
include-package-data = true

[tool.setuptools.packages.find]
include = ["core*", "tools*", "pomera*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.md", "*.ico", "*.png"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.setuptools_scm]
# Write version to _version.py during build
write_to = "pomera/_version.py"
write_to_template = '# AUTO-GENERATED by setuptools_scm - DO NOT EDIT\n__version__ = "{version}"'
# Fallback if no Git info available
fallback_version = "0.0.0"
# Version scheme
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
